+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - Fixed `@argfiles` not being expanded when specified after a subcommand ([#570](https://github.com/ajalt/clikt/pull/570))
 - Fixed syntax error in generated bash completions when an argument name contained spaces ([#563](https://github.com/ajalt/clikt/pull/563))
+- Support bash completions of file parameters when file names contain spaces or special characters ([#379](https://github.com/ajalt/clikt/pull/379))
 
 ## 5.0.2
 ### Changed